+++ cq/tests/test_storages.py
@@ -1,5 +1,6 @@
 from ..storages import Storage
 from unittest import mock
+import inspect
 import pytest
 
 
@@ -8,7 +9,7 @@ import pytest
 def test_store(handle_event):
     storage = Storage()
     with mock.patch.object(storage, 'create_event', return_value='EVENT') as create_event,\
-         mock.patch.object(storage, 'append') as append:
+            mock.patch.object(storage, 'append', return_value='APPENDED_EVENT') as append:
         storage.store('User', 'Registered', 'aabbcc', {'name': 'joe'})
         create_event.assert_called_once_with(
             id='EVENT_ID',
@@ -20,7 +21,7 @@ def test_store(handle_event):
             revision=1,
         )
         append.assert_called_once_with('EVENT')
-        handle_event.assert_called_once_with('EVENT')
+        handle_event.assert_called_once_with('APPENDED_EVENT')
 
 
 def test_local__append(local_storage):
@@ -80,3 +81,14 @@ def test_local__book_unique_fails_for_duplicate(local_storage):
 
     # first value should not be overridden
     assert local_storage.get_unique('user_email', 'joe@doe.com') == 'JOE_ID'
+
+
+def test_local__iter_all_events(local_storage):
+    all_events = [
+        local_storage.store('User', 'Registered', 'JOE_ID', {'name': 'joe'}, 'TS'),
+        local_storage.store('User', 'Activated', 'JOE_ID', {'name': 'joe'}, 'TS'),
+        local_storage.store('User', 'Registered', 'JANE_ID', {'name': 'jane'}, 'TS'),
+    ]
+
+    assert inspect.isgenerator(local_storage.iter_all_events()) is True
+    assert list(local_storage.iter_all_events()) == all_events